requirements.txt
beautifulsoup4==4.8.0
certifi==2019.6.16
chardet==3.0.4
contractions==0.0.21
cssselect==1.0.3
feedfinder2==0.0.4
feedparser==5.2.1
idna==2.8
jieba3k==0.35.1
lxml==4.3.4
markdown2==2.3.8
newspaper3k==0.2.8
nltk==3.4.4
Pillow==6.1.0
pyahocorasick==1.4.0
pyquery==1.4.0
pyspellchecker==0.5.0
python-dateutil==2.8.0
PyYAML==5.1.1
readtime==1.1.1
requests==2.22.0
requests-file==1.4.3
six==1.12.0
soupsieve==1.9.2
textsearch==0.0.17
tinysegmenter==0.3
tldextract==2.2.1
Unidecode==1.1.1
urllib3==1.25.3
